fix the use of incomplete types for C++20 compatibilities
